Initial commit with a little structure.
We've added the ATProto SDK and set up a really basic setup.
It's all Claude generated, but this is a learning exercise to let's take
a look at it.

BlonkAgent - to be renamed, but basically our Bsky client

POST_NSID - our namespace identifier (What is this? We'll come back to that)

PostManager - our interface to create or retrieve posts

`index.ts` - a super basic page skeleton

With this I guess we can try to get some shit on a page.