prompt 3
Let's get a system set up so that we can have a real webpage for this?

We don't want to clone the look of reddit.
We want this to feel more like the old website del.icio.us with some fun modern twists, I think.
Would you happen to have any memories of del.icio.us from the early 2000s that you can draw from here for inspiration?