prompt 6
its failing to detect emerging vibes and we have no server logs indicating this.

that is troublesome.

we need to hash out if this firehose is even working and you are just making willy error handlers that cover important flaws in the system.

Why is it using a search to find the vibes? We should be consuming the entire firehose!