prompt 6 or 7
I thought of a critical flaw. Our posts that we are looking for replies to are totally fresh. We should actually
be sweeping chronologically to check these posts and then enrich them in a processed manner. Let's start building
out that pipeline. It will be a GenServer called HotPostSweeper that sweeps out our hot posts. We will need to
make a hot_posts table and schema, and then save all the metadata needed for them (the entire record) if they do
in fact have a link. Every 10th one. And then every 10 minutes we grab 25 links and check the replies. A link gets
checked for replies 10 times total. The checks are counted on the record. The sweeper will also auto-delete any
link that has hit 10 checks or is older than 1 day. This ensures no weird table growth there.

If we find a post with 5 or more replies, then we will have the sweeper create a blip on the hot_bsky vibe.

Does this make sense?